% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_psofile.R
\name{make_psofile}
\alias{make_psofile}
\title{Write a pso (\pkg{PSObject}) file}
\usage{
make_psofile(PSobject, filename = NULL)
}
\arguments{
\item{PSobject}{an object of class \code{PSobject}}

\item{filename}{(optional) string. If not \code{NULL}, the pso file will be saved via
\code{\link{writeLines}} in a text file named after the string supplied.}
}
\value{
A list of strings, representing a file.
}
\description{
Generate a text file that includes essential information, 
which is read by \code{to_stan()} and converted to a stan file for sampling.
}
