% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_standata.R
\name{make_standata}
\alias{make_standata}
\title{Data for \pkg{PStrata} Models}
\usage{
make_standata(PSobject)
}
\arguments{
\item{PSobject}{an object of class \code{PSObject}}
}
\value{
a named list of objects containing the required data to fit a
\pkg{PStrata} model with \bold{Stan}.
}
\description{
Generate data for \pkg{PStrata} models to be passed to \bold{Stan}
}
