% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse.formula}
\alias{parse.formula}
\title{Extract important information from an R formula object}
\usage{
parse.formula(formula, data)
}
\arguments{
\item{formula}{a formula, syntaxed as \code{lm()} or \code{lme4::lmer()}.}

\item{data}{a data frame, including variables specified in \code{formula}.}
}
\value{
A list, containing at least the following components:
\describe{
\item{formula}{formula as the input argument}
\item{non_random_formula}{formula with only fixed effects}
\item{random_effect}{boolean value of whether a random effect is included in the formula}
\item{response}{symbols that appear on the left hand side of the formula}
\item{has_intercept}{boolean value of whether intercept is included in the formula}
\item{predictors}{symbols of all the linear predictors}
\item{num_of_predictors}{number of linear predictors}
\item{model_matrix}{the design matrix given by the fixed effects of the model}
}

When random effects are included, the following item is also included:
\describe{
\item{random_effect_list}{a list of all the random effects, each of which is a list with a design matrix for the linear terms specific to the effect, the name of these linear terms, and the levels that the random effect can be}
}
}
\description{
Parse the formula to obtain the names of the responses and covariates, and the design matrix
with \code{data} provided. When the formula includes random effects, specified using the syntax
for package \code{lme4}, the design matrix corresponding for each cluster variable is also provided.
}
\seealso{
\code{something}
}
