% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior_flat}
\alias{prior_normal}
\alias{prior_t}
\alias{prior_cauchy}
\alias{prior_lasso}
\alias{prior_logistic}
\alias{prior_chisq}
\alias{prior_inv_chisq}
\alias{prior_exponential}
\alias{prior_gamma}
\alias{prior_inv_gamma}
\alias{prior_weibull}
\title{Prior functions}
\usage{
prior_flat()

prior_normal(mu = 0, sigma = 1)

prior_t(mu = 0, sigma = 1, df = 1)

prior_cauchy(mu = 0, sigma = 1)

prior_lasso(mu = 0, sigma = 1)

prior_logistic(mu = 0, sigma = 1)

prior_chisq(df = 1)

prior_inv_chisq(df = 1)

prior_exponential(beta = 1)

prior_gamma(alpha = 1, beta = 1)

prior_inv_gamma(alpha = 1, beta = 1)

prior_weibull(alpha = 1, sigma = 1)
}
\arguments{
\item{mu, sigma, df, alpha, beta}{parameters for the prior distribution}
}
\value{
A list, including the following items.
\describe{
\item{name}{name of the distribution}
\item{type}{type of the distribution, one character string of "real" or "positive"}
\item{args}{a named list of all the input parameters}
\item{call}{a function call object of the prior distribution on the parameters}
}
}
\description{
Define prior functions used in \code{PStrata}.
}
