% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior_flat}
\alias{prior_flat}
\title{Prior functions}
\usage{
prior_flat()
}
\arguments{
\item{...}{parameters for the prior distribution}
}
\value{
A list, including the following items.
\describe{
\item{name}{name of the distribution}
\item{type}{type of the distribution, one character string of "real" or "positive"}
\item{args}{a named list of all the input parameters}
\item{call}{a function call object of the prior distribution on the parameters}
}
}
\description{
Define prior functions used in \code{PStrata}.
}
