% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOutcome.R
\name{PSOutcome}
\alias{PSOutcome}
\title{Estimated potential outcome for principal stratification analysis}
\usage{
PSOutcome(PStrataObj)
}
\arguments{
\item{PStrataObj}{an object of class \code{PStrata} or \code{PStrata_survival}}
}
\description{
Create an object useful to present the potential outcomes under each treatment arm
for each principal stratum. Contrasts between treatment arms or principal strata
are easy to obtain from this object.
}
