% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{stancode.PStrataFit}
\alias{stancode.PStrataFit}
\title{Extract the Generated Stan Code}
\usage{
stancode.PStrataFit(object, ...)
}
\arguments{
\item{object}{A \code{PStrataFit} object.}

\item{...}{Currently unused.}
}
\value{
The Stan code as a character string.
}
\description{
Extract the Generated Stan Code
}
