% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{The family function for survival data}
\usage{
survival(method = "Cox", link = "identity")
}
\arguments{
\item{method}{the parametric method used for survival data. Can be Cox or AFT.}

\item{link}{a link function, currently only identity is implemented and used}
}
\value{
A \code{family} object
}
\description{
Construct a family object for survival data
}
