% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_txt.R
\name{write.txt}
\alias{write.txt}
\title{Write a PSObject(.pso) file}
\usage{
write.txt(
  S.formula,
  Y.formula,
  Y.family,
  strata,
  ER,
  prior_intercept = prior_flat(),
  prior_coefficient = prior_normal(),
  prior_sigma = prior_inv_gamma(),
  prior_alpha = prior_inv_gamma(),
  prior_lambda = prior_inv_gamma(),
  prior_theta = prior_normal(),
  survival.time.points = 50,
  filename = NULL
)
}
\arguments{
\item{S.formula}{a formula for the principal strata model}

\item{Y.formula}{a formula for the outcome model}

\item{Y.family}{a family object, indicating the family of the model}

\item{strata}{a vector of integers, each integer referring to a stratum that is included in the model}

\item{ER}{a vector of integers, indicating the strata on which the exclusion restriction assumption is assumed.}

\item{survival.time.points}{integer, the number of points to evaluate.}

\item{prior_...}{the prior distribution for the corresponding parameters.}
}
\value{
A text file, readable by \code{to_stan()}.
}
\description{
Generate a text file that includes essential information, which is read by \code{to_stan()} and converted to a stan file for sampling.
}
\seealso{
\code{PStrata}
}
