% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stancode.R
\name{make_stancode}
\alias{make_stancode}
\title{\bold{Stan} Code for \pkg{PStrata} Models}
\usage{
make_stancode(PSobject, filename = NULL)
}
\arguments{
\item{PSobject}{an object of class \code{PSobject}}

\item{filename}{(optional) string. If not \code{NULL}, the stan file will be saved via
\code{\link{cat}} in a text file named after the string supplied.}
}
\value{
A string, which can be printed on screen using \code{\link{cat}}.
}
\description{
Generate the \bold{Stan} code corresponding to the model,
which is read by \bold{Stan} to do sampling.
}
