% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSample.R
\name{PSSample}
\alias{PSSample}
\title{Sample with \code{Stan}}
\usage{
PSSample(
  file,
  model_name = "anon_model",
  model_code = "",
  fit = NA,
  data = list(),
  pars = NA,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  init = "random",
  seed = sample.int(.Machine$integer.max, 1),
  algorithm = c("NUTS", "HMC", "Fixed_param"),
  control = NULL,
  sample_file = NULL,
  diagnostic_file = NULL,
  save_dso = TRUE,
  verbose = FALSE,
  include = TRUE,
  cores = getOption("mc.cores", 1L),
  open_progress = interactive() && !isatty(stdout()) &&
    !identical(Sys.getenv("RSTUDIO"), "1"),
  ...,
  boost_lib = NULL,
  eigen_lib = NULL
)
}
\description{
Sample from the posterior distribution by calling \code{\link[rstan:stan]{stan}}. 
Check \code{\link[rstan:stan]{stan}} for details of the arguments.
}
