% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PStrata.R
\name{PStrata}
\alias{PStrata}
\title{Principal strata analysis for data with post-randomization intervention}
\usage{
PStrata(
  S.formula,
  Y.formula,
  Y.family,
  data,
  strata,
  ER,
  prior_intercept = prior_flat(),
  prior_coefficient = prior_normal(),
  prior_sigma = prior_inv_gamma(),
  prior_alpha = prior_inv_gamma(),
  prior_lambda = prior_inv_gamma(),
  prior_theta = prior_normal(),
  filename = "unnamed",
  ...
)
}
\arguments{
\item{S.formula}{a formula for the principal strata model, in the form of \code{Z + D1 + D2 ~ X1 + X2}, where \code{Z} is the treatment assignment and \code{D1} and \code{D2} are the post-randomization variables.}

\item{Y.formula}{a formula for the outcome model, in the form of \code{Y ~ X1 + X2}, where \code{Y} is the outcome variable}

\item{Y.family}{a family object, indicating the family of the model as used in \code{lm()}.}

\item{data}{a data frame object, the data used to do inference on.}

\item{strata}{a vector of integers, each integer referring to a stratum that is included in the model}

\item{ER}{a vector of integers, indicating the strata on which the exclusion restriction assumption is assumed.}

\item{prior_...}{the prior distribution for the corresponding parameters.}
}
\value{
A text file, readable by \code{to_stan()}.
}
\description{
description
}
